refactor: split top monolithic script into focused modules
Split body_model.py (510 LOC) into a thin facade (69 LOC) plus three
focused sub-modules: axial_skeleton.py, upper_limbs.py, lower_limbs.py.
All public API preserved via re-exports. Sub-modules delegate shared
limb assembly logic to the existing body_helpers.py.

Fixes #123
